work on rendering filtrated cars
