+++ src/scripts/copmponent/CarCard/helper.ts
