+++ src/scripts/copmponent/TabsSection/TabsSection.ts
@@ -1,27 +1,25 @@
 import '../../../scss/component/TabsSection.scss';
-import { TabComponent } from "../TabComponent/TabComponent";
-import CarCardComponent from "../CarCardComponent/CarCardComponent";
-import { CarCardProps, carSetInterface } from "../CarCardComponent/types";
+import { Tab } from "../Tab/Tab.ts";
+import CarCardComponent from "../CarCard/CarCard";
+import { CarCardPropsExpanded, carSetInterface } from "../CarCard/types";
 import Loader from "../Loader/Loader";
-import { GenerateDataToRender } from "../CarCardComponent/helper";
+import { GenerateDataToRender } from "../CarCard/helper";
 
 
 
 const loader = new Loader(document.querySelector('.popular__cars__container') as HTMLElement);
 
 export default class TabsSection {
-  public tabComponent: TabComponent;
-  // public carCardComponent: CarCardComponent;
+  public tab: Tab;
   public carTypeWrapperArray: HTMLElement[] = [];
   public activeTab: HTMLElement;
   public activeCarTypeByDefault: string;
-  public defaultCarDataToRender: CarCardProps[] = [];
+  public defaultCarDataToRender: CarCardPropsExpanded[] = [];
 
   constructor(public tabParams: any, public carDataLink: string | carSetInterface, public linkToPhotoFetch: string, public type: string = "general")  {
     this.carDataLink = carDataLink;
-    this.tabComponent = new TabComponent(this.tabParams, document.querySelector('.popular__tabs__container') as HTMLDivElement);
+    this.tab = new Tab(this.tabParams, document.querySelector('.popular__tabs__container') as HTMLDivElement);
     this.defaultCarDataToRender = []
-    // this.carCardComponent = new CarCardComponent(this.defaultCarDataToRender, document.createElement('div') as HTMLDivElement);
     this.activeTab = document.querySelector('.tab--active') as HTMLElement;
     this.activeCarTypeByDefault = this.activeTab.innerText;
     switch (this.type) {
@@ -40,7 +38,7 @@ export default class TabsSection {
       try {
         this.defaultCarDataToRender = await GenerateDataToRender(this.activeCarTypeByDefault, carDataLink, linkToPhotoFetch)
         loader.remove();
-        this.defaultCarDataToRender.forEach((car: CarCardProps) => {
+        this.defaultCarDataToRender.forEach((car: CarCardPropsExpanded) => {
         new CarCardComponent(car, document.querySelector(`.popular__cars__container`) as HTMLDivElement);
       });
     } catch (error) {
@@ -54,7 +52,7 @@ export default class TabsSection {
       this.defaultCarDataToRender = await GenerateDataToRender(this.activeCarTypeByDefault, this.carDataLink, this.linkToPhotoFetch, 'accordingFilterRequest');
       loader?.remove();
       // console.log(this.defaultCarDataToRender)
-      this.defaultCarDataToRender.forEach((car: CarCardProps) => {
+      this.defaultCarDataToRender.forEach((car: CarCardPropsExpanded) => {
         new CarCardComponent(car, document.querySelector(`.popular__cars__container`) as HTMLDivElement);
       });
     } catch (error) {
@@ -83,7 +81,7 @@ export default class TabsSection {
               try {
                   this.defaultCarDataToRender = await GenerateDataToRender(this.activeTab.innerText, '../../../../dataJSON/carData.json', 'https://api.thecatapi.com/v1/images/search?limit=1');
                   loader.remove();
-                  this.defaultCarDataToRender.forEach((car: CarCardProps) => {
+                  this.defaultCarDataToRender.forEach((car: CarCardPropsExpanded) => {
                       new CarCardComponent(car, document.querySelector('.popular__cars__container') as HTMLDivElement);
                   });
               } catch (error) {
