+++ src/scripts/copmponent/index.ts
@@ -1,2 +1,2 @@
 export { default as Header } from "./Header/Header";
-export { default as CarCardComponent } from "./CarCardComponent/CarCardComponent";
+export { default as CarCardComponent } from "./CarCard/CarCard";
