+++ src/scripts/main.ts
@@ -1,7 +1,7 @@
 import "../scss/main.scss";
 import fetchData from "./copmponent/Select/helpers";
 import TabsSection from "./copmponent/TabsSection/TabsSection";
-import { tabParams } from "./copmponent/TabComponent/TabComponent";
+import { tabParams } from "./copmponent/Tab/Tab";
 import renderTestimonials from "./copmponent/Testimonials/renderTestimonials";
