+++ src/scss/component/TabsSection.scss
@@ -5,7 +5,7 @@
 .popular__tabs__container{
   width: 100%;
   display: grid;
-  grid-template-columns: repeat(4, 1fr);
+  grid-template-columns: repeat(5, 1fr);
   margin-top: 50px
 }
