+++ src/scss/main.scss
@@ -4,7 +4,7 @@
 @import "./helper/fonts.scss";
 @import "./helper/layout.scss";
 
-@import "./component/carCardComponent.scss";
+@import "./component/carCard.scss";
 @import "./component/TabsSection.scss";
 @import "./component/testimonials.scss";
 @import "./component/footer.scss";
